Move schedulers to their own directory.
