+++ tools/build_variables.bzl
@@ -410,8 +410,8 @@ libtorch_cuda_core_sources = [
     "torch/csrc/jit/codegen/cuda/predicate_compute.cpp",
     "torch/csrc/jit/codegen/cuda/register_interface.cpp",
     "torch/csrc/jit/codegen/cuda/root_domain_map.cpp",
-    "torch/csrc/jit/codegen/cuda/scheduler_registry.cpp",
-    "torch/csrc/jit/codegen/cuda/scheduler.cpp",
+    "torch/csrc/jit/codegen/cuda/scheduler/registry.cpp",
+    "torch/csrc/jit/codegen/cuda/scheduler/scheduler.cpp",
     "torch/csrc/jit/codegen/cuda/shape_inference.cpp",
     "torch/csrc/jit/codegen/cuda/fusion_segmenter.cpp",
     "torch/csrc/jit/codegen/cuda/tensor_view.cpp",
